web/requirements.txt
Django==1.7
django-crispy-forms==1.5.2
django-filter==0.11.0
django-registration-redux==1.2
djangorestframework==3.2.4
Markdown==2.6.2
Pillow==3.0.0
wheel==0.24.0
